use round robin when connect to public and then private proxies
and connect to them in download_clients/threads upon usage, rather than on start
If there is only one file to download, it will make only one connection (so faster)
downside it is that we can see same message up to 8 (threads) times (if server is not available).
Just use round robin with shuffled start.