otp command for client
this basically allows to not sharing secrets on proxies, especially on local proxies (to be done)
on (at least public) proxies we may still probably share explicit secret, rather than giving access to CVS.
Changing secrets is easy and should be done anyway in maintaining. And so no CVS access is needed.

but on local user PCs sharing secret is dangerous, and allowing to only work with private proxies is limiting for all remote workers